fix(config): Yarn PnP Support
