+++ .changeset/tiny-trees-doubt.md
@@ -1,0 +1,7 @@
+---
+'@graphql-mesh/cli': patch
+'@graphql-mesh/config': patch
+'@graphql-mesh/utils': patch
+---
+
+fix(config): Yarn PnP Support
